Add sample filtering
A new sample will be sent to Carbon only if its value is different that
the previous value. Pattern matching is used to test for equivalence
(i.e. =:= not ==).

Some check on timestamp could be added so that historical samples
are forwarded no matter what the value is.
That could have a use case when correcting data points with errors
or filling holes in the historical data. TBD in the future.